add permission to controllers

+++ Modules/Post/Http/Controllers/CategoryController.php
@@ -9,6 +9,8 @@ use Illuminate\Support\Facades\Session;
 use Illuminate\Validation\Validator;
 use Modules\Post\Entities\Category;
 use Modules\Post\Entities\Language;
+use Spatie\Permission\Traits\HasRoles;
+use Spatie\Permission\Traits\HasPermissions;
 
 class CategoryController extends Controller
 {
@@ -19,10 +21,15 @@ class CategoryController extends Controller
 
     public function index()
     {
-        $languages = Language::all();
-        $categories = Category::with('lang')->with('parent')->get();
+        if(auth()->user()->hasRole('admin') || auth()->user()->hasPermissionTo('read category')) {
+            $languages = Language::all();
+            $categories = Category::with('lang')->with('parent')->get();
 
-        return view('post::category' , compact('languages','categories'));
+            return view('post::category', compact('languages', 'categories'));
+        }
+        else{
+            return view('layouts.error.404');
+        }
     }
 
     public function catsBylang(Request $request)
@@ -34,46 +41,56 @@ class CategoryController extends Controller
 
     public function store(Request $request)
     {
-        $request->validate([
-            'title'    => 'required',
-            'language' => 'required',
-        ]);
+        if(auth()->user()->hasRole('admin') || auth()->user()->hasPermissionTo('create category')) {
+            $request->validate([
+                'title' => 'required',
+                'language' => 'required',
+            ]);
 
-        $parent = $request->parent;
-        if($request->parent == ""){
-            $parent = null;
+            $parent = $request->parent;
+            if ($request->parent == "") {
+                $parent = null;
+            }
+            $category = new Category();
+            $category->title = $request->title;
+            $category->lang_id = $request->language;
+            $category->parent_id = $parent;
+            $category->save();
+
+            Session::flash('success', 'New Category was added successfuly');
+
+            return back();
         }
-        $category = new Category();
-        $category->title = $request->title;
-        $category->lang_id = $request->language;
-        $category->parent_id = $parent;
-        $category->save();
-
-        Session::flash('success' , 'New Category was added successfuly');
-
-        return back();
+        else{
+            return view('layouts.error.403');
+        }
     }
 
     public function update($id,Request $request)
     {
-        $request->validate([
-            'title'    => 'required',
-            'language' => 'required',
-        ]);
+        if(auth()->user()->hasRole('admin') || auth()->user()->hasPermissionTo('update category')) {
+            $request->validate([
+                'title' => 'required',
+                'language' => 'required',
+            ]);
 
-        $parent = $request->parent;
-        if($request->parent == ""){
-            $parent = null;
+            $parent = $request->parent;
+            if ($request->parent == "") {
+                $parent = null;
+            }
+
+            $cat = Category::find($id);
+            $cat->title = $request->title;
+            $cat->lang_id = $request->language;
+            $cat->parent_id = $parent;
+            $cat->save();
+
+            Session::flash('success', 'Category was Updated successfuly');
+            return back();
         }
-
-        $cat = Category::find($id);
-        $cat->title = $request->title;
-        $cat->lang_id = $request->language;
-        $cat->parent_id = $parent;
-        $cat->save();
-
-        Session::flash('success' , 'Category was Updated successfuly');
-        return back();
+    else{
+        return view('layouts.error.403');
+    }
 
 
     }
@@ -84,12 +101,17 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
-        $category = Category::find($id);
-        $category->posts()->sync([]);
-        $category->childs()->update(['parent_id' => null]);
-        $category->delete();
+        if(auth()->user()->hasRole('admin') || auth()->user()->hasPermissionTo('destroy category')) {
+            $category = Category::find($id);
+            $category->posts()->sync([]);
+            $category->childs()->update(['parent_id' => null]);
+            $category->delete();
 
-        Session::flash('delete' , 'Category was deleted successfuly');
-        return back();
+            Session::flash('delete', 'Category was deleted successfuly');
+            return back();
+        }
+        else{
+            return view('layouts.error.403');
+        }
     }
 }
